Improve the band-based correllator
We went back to sliding track, but instead of 160 tracks we have 2.
This is because we now decimate first, and averate over a half-bit
after that.

The current version has a good functional performance, but not
a good performance regarding the efficiency: burns 62% CPU
where rtl_adsb burns 9.3%. Of course, we consume 20 ms/s and
rtl_adsb consumes 2 ms/s, so we are actually more efficient
per sample. But we decimate by 5, so there's actually plenty
to improve. This is essential for a future port to Raspberry Pi.